Corrected spawn rate, added basic visual cues.
This is a pretty unforgiving game!

Notes now spawn based on a file loaded in Track. Thanks to Robbie's
correction of my dorkish code it seems to be responding better!

Notes will fade out over a few seconds after they are pressed.

You can hit a note a) far too early (which counts as a failure and
causes the note to disappear) b) badly (fail), c) with good/great
accuracy (which means a success).

Colours are mostly for debugging.

Next steps (in no particular order)
- Play a sound on press (better form of feedback than colour-changing,
since we want the player to use their ears)
- Tidy up after ourselves once the battle is done.
- Have some kind of visual indicator to show where the note must be to
get a good/great score.
- Support multiple button types.